Updated project with errors and exceptions handling code

+++ Bakend_Spring_Project/lvc_spring_backend/src/main/java/com/lvc/app/lvc_spring_backend/services/UserServices.java
@@ -10,6 +10,8 @@ import org.springframework.stereotype.Service;
 
 import com.lvc.app.lvc_spring_backend.dos.UserDO;
 import com.lvc.app.lvc_spring_backend.dtos.UserDTO;
+import com.lvc.app.lvc_spring_backend.exception.ResourceAlreadyExistsException;
+import com.lvc.app.lvc_spring_backend.exception.ResourceNotFoundException;
 import com.lvc.app.lvc_spring_backend.repository.UserRepository;
 
 @Service
@@ -23,6 +25,8 @@ public class UserServices {
 
 	public UserDTO add(UserDTO user) {
 		UserDO userDO = modelMapper.map(user, UserDO.class);
+		if(userRepository.existsByName(userDO.getName()))
+			throw new ResourceAlreadyExistsException("User with name " + user.getName() + " already exists");
 		return modelMapper.map(userRepository.save(userDO), new TypeToken<UserDTO>() {
 		}.getType());
 	}
@@ -33,11 +37,17 @@ public class UserServices {
 	}
 
 	public UserDTO getById(Long id) {
+		Optional<UserDO> user = userRepository.findById(id);
+		if(!user.isPresent())
+			throw new ResourceNotFoundException("User with id " + id + " not exists");
 		return modelMapper.map(userRepository.findById(id).get(), new TypeToken<UserDTO>() {
 		}.getType());
 	}
 
 	public List<UserDTO> remove(Long id) {
+		Optional<UserDO> user = userRepository.findById(id);
+		if(!user.isPresent())
+			throw new ResourceNotFoundException("User with id " + id + " not exists");
 		userRepository.deleteById(id);
 		return fetchAll();
 	}
